create and delete cluster

+++ pkg/create_alluxio_cluster.go
@@ -4,46 +4,44 @@ import (
 	"fmt"
 	"github.com/Alluxio/k8s-operator/pkg/logger"
 	"os"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/yaml"
 )
 
 const chartPath = "/opt/alluxio-helm-chart"
 
-func CreateAlluxioCluster(ctx ReconcileRequestContext) ReconcileResponse {
-	logger.Infof("Creating Alluxio cluster %v in namespace %v.", ctx.Name, ctx.Namespace)
-
-	// Construct config.yaml file
-	data, err := yaml.Marshal(ctx.AlluxioCluster)
+func CreateAlluxioClusterIfNotExist(ctx ReconcileRequestContext) error {
+	// if the release has already been deployed, requeue without further actions
+	helmCtx := HelmContext{
+		Namespace:   ctx.Namespace,
+		ReleaseName: ctx.Name,
+	}
+	exists, err := checkIfHelmReleaseExists(helmCtx)
 	if err != nil {
-		return ReconcileResponse{
-			Err:        err,
-			NeedReturn: true,
-			Result:     ctrl.Result{},
-		}
+		return err
+	}
+	if exists {
+		return nil
+	}
+
+	logger.Infof("Creating Alluxio cluster %v in namespace %v.", ctx.Name, ctx.Namespace)
+	// Construct config.yaml file
+	data, err := yaml.Marshal(ctx.AlluxioCluster.Spec)
+	if err != nil {
+		return err
 	}
 	configYamlFile, err := os.CreateTemp(os.TempDir(), fmt.Sprintf("%s-%s-config.yaml", ctx.Namespace, ctx.Name))
 	configYamlFilePath := configYamlFile.Name()
 	if err != nil {
 		logger.Errorf("failed to create empty config file: %v", err)
-		return ReconcileResponse{
-			Err:        err,
-			NeedReturn: true,
-			Result:     ctrl.Result{},
-		}
+		return err
 	}
 	err = os.WriteFile(configYamlFilePath, data, 0400)
 	if err != nil {
 		logger.Errorf("failed saving config file: %v", err)
-		return ReconcileResponse{
-			Err:        err,
-			NeedReturn: true,
-			Result:     ctrl.Result{},
-		}
+		return err
 	}
-
 	// helm install release with the constructed config.yaml
-	helmCtx := HelmContext{
+	helmCtx = HelmContext{
 		HelmChartPath:      chartPath,
 		ConfigFilePath:     configYamlFilePath,
 		Namespace:          ctx.Namespace,
@@ -54,16 +52,8 @@ func CreateAlluxioCluster(ctx ReconcileRequestContext) ReconcileResponse {
 		logger.Errorf("error installing helm release. Uninstalling...")
 		if err := HelmDeleteIfExist(helmCtx); err != nil {
 			logger.Errorf("failed to delete failed helm release %v in namespace %v: %v", ctx.Name, ctx.Namespace, err)
-			return ReconcileResponse{
-				Err:        err,
-				NeedReturn: true,
-				Result:     ctrl.Result{},
-			}
+			return err
 		}
 	}
-	return ReconcileResponse{
-		Err:        nil,
-		NeedReturn: true,
-		Result:     ctrl.Result{Requeue: true},
-	}
+	return nil
 }
